requirements.txt
# requirements.txt

python-dotenv
requests
pyttsx3
SpeechRecognition
sounddevice
soundfile
numpy
webrtcvad
psutil
Pillow